backend/requirements.txt
fastapi
uvicorn
python-dotenv
Pillow
torch
transformers
google-genai
firebase-admin
requests